% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_day}
\alias{count_day}
\title{Count of judgements by Day}
\format{A data frame with 5790 rows and 2 variables:
\describe{
  \item{judgmentDate}{Date of judgement}
  \item{count}{Number of judgements}
}}
\usage{
count_day
}
\description{
Count of judgements by Day
}
\keyword{datasets}
