% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_week_day}
\alias{count_week_day}
\title{Count of judgements by Day of the Week}
\format{A data frame with 7 rows and 2 variables:
\describe{
  \item{weekDay}{Day of the week}
  \item{count}{Number of judgements}
}}
\usage{
count_week_day
}
\description{
Count of judgements by Day of the Week
}
\keyword{datasets}
