% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{count_year_month}
\alias{count_year_month}
\title{Count of judgements by Year and Month}
\format{A data frame with 216 rows and 2 variables:
\describe{
  \item{yearMonth}{Year and month}
  \item{count}{Number of judgements}
}}
\usage{
count_year_month
}
\description{
Count of judgements by Year and Month
}
\keyword{datasets}
