% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exp_loss.R
\name{exp_loss}
\alias{exp_loss}
\title{Expected value of loss amount per business day.}
\usage{
exp_loss(average.consult, prob.hire, average.ticket, language = "en")
}
\arguments{
\item{average.consult}{Average number of mensal credit bureau consultations.}

\item{prob.hire}{(Historical) probability of hiring. Can be considered as the number of converted customers divided by the number of  credit bureau consultations in a period of time.}

\item{average.ticket}{The average ticket of regular customers.}
}
\value{
\code{$expected.value} Expected value of loss amount per business day of negative credit bureau attribution.

\code{$text} An automatic text that details the \code{$expected.value}.
}
\description{
Expected value of loss amount per business day.
}
\examples{
library(jurimetrics)
exp_loss(1000, 0.1, 3450)
}
\references{
Feller, W. (1957) \href{https://archive.org/details/AnIntroductionToProbabilityTheoryAndItsApplicationsVolume1}{An Introduction to Probability Theory and its Applications}.
}
