% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pattern_danomoral.R
\name{pattern_danomoral}
\alias{pattern_danomoral}
\title{Pattern of 'dano moral'}
\usage{
pattern_danomoral()
}
\value{
A char vector listing the 'dano moral' pattern.
}
\description{
Pattern of 'dano moral'
}
\examples{
pattern_danomoral()
}
