% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poisson1837}
\alias{poisson1837}
\title{Poisson Jurimetrics Data}
\format{A data frame with 24 rows and 5 variables:
\describe{
  \item{year}{Year}
  \item{accused}{Number of accusations}
  \item{convicted}{Number of convicted}
  \item{rate}{Conviction rate}
  \item{group}{Region}
}}
\usage{
poisson1837
}
\description{
Poisson Jurimetrics Data
}
\keyword{datasets}
