% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_inj.R
\name{pre_inj}
\alias{pre_inj}
\title{Expected value of loss amount per business day of mistaken credit bureau attribution.}
\usage{
pre_inj(average.consult, prob.hire, average.ticket)
}
\arguments{
\item{average.consult}{Average number of mensal credit bureau consultations.}

\item{prob.hire}{(Historical) probability of hiring. Can be considered as the number of converted customers divided by the number of  credit bureau consultations in a period of time.}

\item{average.ticket}{The average ticket of regular customers.}
}
\value{
\code{$fcast} predicted time series using the model that minimizes the forecasting mean square error.
\code{$runtime} running time.
\code{mse.pred} mean squared error of prediction. Used to decide the best model.
}
\description{
Expected value of loss amount per business day of mistaken credit bureau attribution.
}
\examples{
library(jurimetrics)

pre_inj(1000, 0.1, 3450)
}
\references{
Feller, W. (1957) An Introduction to Probability Theory and its Applications. https://archive.org/details/AnIntroductionToProbabilityTheoryAndItsApplicationsVolume1
}
