% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fits.R
\name{fits}
\alias{fits}
\title{Fits the best model from classes ARIMA, ETS, TBATS and NNETAR.}
\usage{
fits(x, train = 0.8, steps = NULL, max.points = 500,
  show.main.graph = T, show.sec.graph = F, PI = F)
}
\arguments{
\item{x}{A vector or ts object.}

\item{train}{The (initial) percentage of the time series to be used to train the models. Must be \code{0 < train < 1}.}

\item{steps}{Number of steps to forecast. If \code{NULL}, uses the number of points not used in training (testing points). Can't be less than the number of testing points.}

\item{max.points}{Limits the maximum number of points to be used in modeling. Uses the last \code{max.points} points of the series.}

\item{show.main.graph}{Logical. Should the main graphic (with the final model) be displayed?}

\item{show.sec.graph}{Logical. Should the secondary graphics (with the training models) be displayed?}

\item{PI}{Prediction Interval used in nnar models. Must take long time processing.}
}
\value{
The predicted time series using the model that minimizes the forecasting mean square error.
}
\description{
Fits the best model from classes ARIMA, ETS, TBATS and NNETAR.
}
\examples{
library(jurimetrics)
library(fpp2)

fits(livestock)
fits(livestock, graf = F)
fits(h02, .9)
fits(gas)

data(count_year_month)
y <- ts(count_year_month$count, start = c(2000,1), frequency = 12)
fits(y, train = 0.8, steps = 24)
}
\references{
Hyndman, R.J., & Athanasopoulos, G. (2018) Forecasting: principles and practice, 2nd edition, OTexts: Melbourne, Australia. https://otexts.com/fpp2.

Zabala, F. J. and Silveira, F. F. (2019). Decades of Jurimetrics. arXiv.org...
}
