% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_doj.R
\name{theme_doj}
\alias{theme_doj}
\title{Defines the Decades of Jurimetrics theme}
\usage{
theme_doj()
}
\value{
Decades of Jurimetrics theme, based on ggplot2::theme_dark.
}
\description{
Defines the Decades of Jurimetrics theme
}
\examples{
library(ggplot2)

theme_doj()
body(theme_doj)
}
\references{
Wickham, H. ggplot2: Elegant Graphics for Data Analysis. Springer-Verlag New York, 2016.

Zabala, F. J. and Silveira, F. F. (2019). Decades of Jurimetrics. arXiv.org...
}
